pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mycelium"
version = "0.1.0"
description = "Mycelium Agent Framework - 从一只蜜蜂到蜂群"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
